Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/user-activation/propagation-sameorigin.tentative-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
User activation propagation across same-origin frame boundary

Tests that user activation propagates across same-origin frame boundary.

Click anywhere on the green area (child frame).


Harness Error (TIMEOUT), message = null

NOTRUN Propagation test
